# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Multiple Choice Question Platform API.\nThis application supports the creation, managing, marking, viewing, retrieving stats, and scores of quizzes.",
        "title": "Multiple Choice Question Platform.",
        "contact": {
            "name": "Saad Ur Rahman",
            "url": "https://www.linkedin.com/in/saad-ur-rahman/",
            "email": "saad.ur.rahman@gmail.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://opensource.org/licenses/GPL-3.0"
        },
        "version": "1.0.1"
    },
    "host": "localhost:44243",
    "basePath": "/api/rest/v1",
    "paths": {
...
        "/user/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a user stored in the database by marking it as deleted. The user must supply their login credentials as well as complete the following confirmation message: \"I understand the consequences, delete my user account USERNAME HERE\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user users delete security"
                ],
                "summary": "Deletes a user. The user must supply their credentials as well as a confirmation message.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "description": "The request payload for deleting an account",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_rest.DeleteUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message with a confirmation of a deleted user account",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Success"
                        }
                    },
                    "400": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    },
                    "409": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    },
                    "500": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Logs in a user by validating credentials and returning a JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user users login security"
                ],
                "summary": "Login a user.",
                "operationId": "loginUser",
                "parameters": [
                    {
                        "description": "Username and password to login with",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_cassandra.UserLoginCredentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT in the api-key",
                        "schema": {
                            "$ref": "#/definitions/model_rest.JWTAuthResponse"
                        }
                    },
                    "400": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    },
                    "403": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    },
                    "500": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    }
                }
            }
        },
        "/user/refresh": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Refreshes a user's JWT by validating it and then issuing a fresh JWT with an extended validity time. JWT must be expiring in under 60 seconds.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user users login refresh security"
                ],
                "summary": "Refresh a user's JWT by extending its expiration time.",
                "operationId": "loginRefresh",
                "responses": {
                    "200": {
                        "description": "A new valid JWT",
                        "schema": {
                            "$ref": "#/definitions/model_rest.JWTAuthResponse"
                        }
                    },
                    "400": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    },
                    "403": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    },
                    "500": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    },
                    "510": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "Creates a user account by inserting credentials into the database. A hashed password is stored.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user users register security"
                ],
                "summary": "Register a user.",
                "operationId": "registerUser",
                "parameters": [
                    {
                        "description": "Username, password, first and last name, email address of user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_cassandra.UserAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "a valid JWT token for the new account",
                        "schema": {
                            "$ref": "#/definitions/model_rest.JWTAuthResponse"
                        }
                    },
                    "400": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    },
                    "409": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    },
                    "500": {
                        "description": "error message with any available details in payload",
                        "schema": {
                            "$ref": "#/definitions/model_rest.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model_cassandra.UserAccount": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "model_cassandra.UserLoginCredentials": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "model_rest.DeleteUserRequest": {
            "type": "object",
            "required": [
                "confirmation",
                "password",
                "username"
            ],
            "properties": {
                "confirmation": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 8
                },
                "username": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "model_rest.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payload": {}
            }
        },
        "model_rest.JWTAuthResponse": {
            "type": "object",
            "required": [
                "expires",
                "threshold",
                "token"
            ],
            "properties": {
                "expires": {
                    "description": "Expiration time as unix time stamp. Strictly used by client to gauge when to refresh the token.",
                    "type": "integer"
                },
                "threshold": {
                    "description": "The window in seconds before expiration during which the token can be refreshed.",
                    "type": "integer"
                },
                "token": {
                    "description": "JWT string sent to and validated by the server.",
                    "type": "string"
                }
            }
        },
        "model_rest.Success": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payload": {}
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}